requirements.txt
torch
pytorch-lightning
transformers
hydra-core
dvc